refactor: move friend_requests to users
