+++ backend/src/app.module.ts
@@ -10,7 +10,7 @@ import { AuthModule } from './auth/auth.module';
 import { TfaModule } from './auth/tfa/tfa.module';
 import { AppConfigModule } from './config/app-config.module';
 import { TypeOrmConfigModule } from './config/typeorm-config.module';
-import { FriendRequestModule } from './friend_requests/friend_request.module';
+import { FriendRequestModule } from './users/friend_requests/friend_request.module';
 import { GuestCleanupService } from './users/guest-cleanup/guest-cleanup.service';
 import { ScheduleModule } from '@nestjs/schedule';
 import { StatusModule } from './users/status/status.module';
