+++ backend/src/users/friend_requests/friend_request.controller.ts
@@ -11,9 +11,9 @@ import {
 } from '@nestjs/common';
 import { FriendRequestService } from './friend_request.service';
 import { CreateFriendRequestDto } from './dto';
-import { AuthenticatedGuard } from '../auth/guards';
-import { GetUser } from '../core/decorators';
-import { User } from '../core/entities/user.entity';
+import { AuthenticatedGuard } from '../../auth/guards';
+import { GetUser } from '../../core/decorators';
+import { User } from '../../core/entities/user.entity';
 
 @UseGuards(AuthenticatedGuard)
 @Controller('friend_requests')
