+++ backend/src/users/friend_requests/friend_request.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { FriendRequestService } from './friend_request.service';
 import { FriendRequestController } from './friend_request.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User, FriendRequest } from '../core/entities';
+import { User, FriendRequest } from '../../core/entities';
 
 @Module({
   imports: [TypeOrmModule.forFeature([FriendRequest, User])],
